Experiment with specialized product schemas
