+++ core/src/main/scala-3/caliban/schema/macros/Macros.scala
@@ -1,6 +1,6 @@
 package caliban.schema.macros
 
-import caliban.schema.Annotations.GQLExcluded
+import caliban.schema.Annotations.{ GQLExcluded, GQLValueType }
 
 import scala.quoted.*
 
@@ -10,6 +10,8 @@ object Macros {
   inline def isFieldExcluded[P, T]: Boolean = ${ isFieldExcludedImpl[P, T] }
   inline def isEnumField[P, T]: Boolean     = ${ isEnumFieldImpl[P, T] }
   inline def implicitExists[T]: Boolean     = ${ implicitExistsImpl[T] }
+  inline def hasAnnotation[T, Ann]: Boolean = ${ hasAnnotationImpl[T, Ann] }
+  inline def hasParams[T]: Boolean          = ${ hasParamsImpl[T] }
 
   /**
    * Tests whether type argument [[FieldT]] in [[Parent]] is annotated with [[GQLExcluded]]
@@ -21,6 +23,16 @@ object Macros {
       fieldName.map(_ == v.name).getOrElse(false)
       && v.annotations.exists(_.tpe =:= TypeRepr.of[GQLExcluded])
     })
+  }
+
+  private def hasParamsImpl[T: Type](using qctx: Quotes): Expr[Boolean] = {
+    import qctx.reflect.*
+    Expr(TypeRepr.of[T].typeSymbol.primaryConstructor.paramSymss.flatten.nonEmpty)
+  }
+
+  private def hasAnnotationImpl[T: Type, Ann: Type](using qctx: Quotes): Expr[Boolean] = {
+    import qctx.reflect.*
+    Expr(TypeRepr.of[T].typeSymbol.annotations.exists(_.tpe =:= TypeRepr.of[Ann]))
   }
 
   private def implicitExistsImpl[T: Type](using q: Quotes): Expr[Boolean] = {
